SpriteSheetFactory loads sprite sheets, TileFactory creates map tiles
SpriteSheetFactory.java: Loads sprite sheets from a specified input,
usable on a per-class, multi-sheet basis. Names each sheet, and
allows loading a region from the sheet.

Done on a per-class basis because this is intended to be used in the
game worlds, where one is loaded at a time, so instances are minimal.

TileFactory.java: Currently proof-of-concept, to show the management of
textures for each different tile. If a snow-tile needs a different
texture but is the same tile, the factory can do it through the
SpriteSheetFactory passed from the world.

Bugs fixed:
Tiles rendered as a tile, then  each one pixel from an edge over. Now
each tile renders correctly by multiplying the coordinate by a factor of
 the tile size.