Refactors MapRenderer to the RenderingSystem and a rendering strategy
Previous MapRenderer would just create the instance of the Tiled map,
and use a custom map renderer to only render the static tiles from the
map.

Now the map is loaded as an entity and assigned its MapComponent
and RenderableComponent, and a rendering strategy MapRenderer is created
to render the map from the RenderableComponent.